api-isolated/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
requests==2.31.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
whitenoise==6.6.0
python-dotenv==1.0.0
supabase==1.0.3
django-cors-headers==4.3.1
PyJWT==2.8.0
pytest==7.4.3
pytest-django==4.7.0
uvicorn==0.27.1 